services/westbank-alerts/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
telethon==1.36.0
aiosqlite==0.20.0
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
python-dotenv==1.0.1
websockets==12.0
beautifulsoup4==4.12.3
